servicios y mejoras

+++ src/main/java/um/programacion2/prestamo/PrestamoService.java
@@ -3,7 +3,7 @@ package um.programacion2.prestamo;
 import java.util.List;
 
 public interface PrestamoService {
-    Prestamo buscarPrestamoPorId(int id);
+    Prestamo buscarPrestamoPorId(Long id);
     List<Prestamo> listarPrestamos();
 
     void solicitarPrestamo(Prestamo prestamo);
